chutamae ya no esta el cabezon

+++ src/main/java/game/entities/Enemy.java
@@ -4,7 +4,7 @@ public abstract class Enemy extends Entity {
     //Atributos
 
     //Constructor
-    public Enemy(double x, int speed, int maxHealth, int health, int damage, double heigth, double width, double y) {
+    public Enemy(double x, double speed, int maxHealth, int health, int damage, double heigth, double width, double y) {
         super(x, speed, maxHealth, health, damage, heigth, width, y);
     }
 }
